fix(conventional-commits): Remove useless checkout
Checking out the repository is not necessary for the
conventional-commits checker action, as evidenced by this repo's own
workflow. Reading the code, it reaches out directly to github.

Reading the code, the code is also *terrible*. Clunky, inflexible,
hacked-together, and barely maintained. I'd like to get off this
third-party dependency ASAP.