chore(docs): Fix TechDocs so that they rendar (#80)
The old config wasn't building the docs because the techdocs annotation was commented out. Now, the docs will correctly build. I also moved the readme under docs/ so we could use github magic to only edit the file in a single spot and still get the nice README in the GitHub UI page