feat: Bump default versions of Python, Node, Ruff, Terraform, Terragrunt (#103)
chore: Update README

The default versions of many of the tools used have started to lag,
so this update just bumps the versions of those tools where
appropriate. They should all still be adjustable on the users'
side if needed. Puppet and Ruby versions were not changed to closer
align to BITS' current Puppet environment.